# src/main/resources/openapi/healthcare-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Healthcare OPD Platform API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/patient/register": {
      "post": {
        "tags": ["Auth"],
        "operationId": "registerPatient",
        "summary": "Register a new patient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PatientRegisterRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Patient registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments/slots": {
      "get": {
        "tags": ["Appointments"],
        "operationId": "getAvailableSlots",
        "summary": "Get available slots for a doctor on a given date",
        "parameters": [
          { "name": "doctorId", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "date", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "Available slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SlotResponse" }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/appointments": {
      "post": {
        "tags": ["Appointments"],
        "operationId": "bookAppointment",
        "summary": "Book an appointment",
        "description": "PATIENT books for self (patientId resolved from JWT if omitted). ADMIN books on behalf of a patient at their hospital.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AppointmentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment booked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AppointmentResponse" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Appointments"],
        "operationId": "getAppointments",
        "summary": "List appointments for the current user",
        "description": "PATIENT sees own appointments. DOCTOR sees own appointments. ADMIN sees appointments for their hospital.",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "date", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated appointment list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AppointmentSummaryPagedResponse" }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{id}": {
      "get": {
        "tags": ["Appointments"],
        "operationId": "getAppointmentById",
        "summary": "Get an appointment by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Appointment detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AppointmentResponse" }
              }
            }
          }
        }
      }
    },
    "/api/appointments/{id}/status": {
      "patch": {
        "tags": ["Appointments"],
        "operationId": "updateAppointmentStatus",
        "summary": "Update the status of an appointment",
        "description": "ADMIN and PATIENT (own appointments) can update status, e.g. cancel or complete.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StatusUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment status updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AppointmentResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/doctors": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "getDoctors",
        "summary": "List doctors, optionally filtered by hospital",
        "parameters": [
          { "name": "hospitalId", "in": "query", "schema": { "type": "integer", "format": "int64" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated doctor list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DoctorSummaryPagedResponse" }
              }
            }
          }
        }
      }
    },
    "/api/doctors/{id}": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "getDoctorById",
        "summary": "Get doctor details by id",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Doctor detail",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DoctorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/doctors/{id}/schedule": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "getDoctorSchedule",
        "summary": "Get a doctor's weekly recurring schedule",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": {
            "description": "Doctor schedule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ScheduleResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Doctors"],
        "operationId": "saveDoctorSchedule",
        "summary": "Save a doctor's weekly recurring schedule",
        "description": "One schedule row per doctor per weekday (day_of_week 0=Mon ... 6=Sun).",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ScheduleRequest" }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule saved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ScheduleResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/doctors/{id}/calendar": {
      "get": {
        "tags": ["Doctors"],
        "operationId": "getDoctorCalendar",
        "summary": "Get a doctor's calendar availability between two dates",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64" } },
          { "name": "from", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "Calendar availability",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CalendarDayResponse" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "type": "object",
        "description": "Authentication response containing the JWT and caller identity information.",
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT token issued after successful authentication.",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIn0.signature"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Authenticated user identifier.",
            "example": 101
          },
          "name": {
            "type": "string",
            "description": "Authenticated user display name.",
            "example": "Dr. Meera Shah"
          },
          "role": {
            "type": "string",
            "description": "Authenticated user role.",
            "enum": [
              "PATIENT",
              "DOCTOR",
              "ADMIN"
            ],
            "example": "DOCTOR"
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Hospital identifier for doctor and admin logins; null for patients.",
            "example": 12
          }
        },
        "required": [
          "token",
          "id",
          "name",
          "role"
        ]
      },
...
      "PatientRegisterRequest": {
        "type": "object",
        "description": "Request payload to register a new patient.",
        "properties": {
          "name": {
            "type": "string",
            "description": "Patient full name.",
            "example": "Asha Patil"
          },
          "gender": {
            "type": "string",
            "description": "Patient gender.",
            "example": "FEMALE"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "description": "Patient date of birth.",
            "example": "1992-08-15"
          },
          "mobileNumber": {
            "type": "string",
            "pattern": "^[0-9]{10}$",
            "description": "Patient 10-digit mobile number.",
            "example": "9876543210"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Patient email address.",
            "example": "asha.patil@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "Patient account password (min 8 characters).",
            "example": "Secret@123"
          },
          "address": {
            "type": "string",
            "description": "Patient residential address.",
            "example": "12 MG Road, Pune, MH"
          },
          "emergencyContactNumber": {
            "type": "string",
            "pattern": "^[0-9]{10}$",
            "description": "Emergency contact 10-digit mobile number.",
            "example": "9876500000"
          },
          "bloodGroup": {
            "type": "string",
            "enum": ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"],
            "description": "Patient blood group.",
            "example": "O+"
          },
          "permanentIllness": {
            "type": "string",
            "description": "Long-term / permanent illness details, if any.",
            "example": "Type 2 Diabetes"
          }
        },
        "required": [
          "name",
          "gender",
          "dob",
          "mobileNumber",
          "password"
        ]
      },
      "AppointmentRequest": {
        "type": "object",
        "description": "Request payload to book an appointment.",
        "properties": {
          "patientId": {
            "type": "integer",
            "format": "int64",
            "description": "Patient identifier. Optional for patient self-booking because it is resolved from the JWT.",
            "example": 201
          },
          "doctorId": {
            "type": "integer",
            "format": "int64",
            "description": "Doctor identifier.",
            "example": 301
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64",
            "description": "Hospital identifier.",
            "example": 12
          },
          "appointmentDate": {
            "type": "string",
            "format": "date",
            "description": "Appointment date.",
            "example": "2026-07-01"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Appointment slot start time in local time.",
            "example": "10:30:00"
          },
          "isOffline": {
            "type": "boolean",
            "default": false,
            "description": "Whether the appointment is booked for an offline visit.",
            "example": false
          },
          "notes": {
            "type": "string",
            "description": "Optional booking notes.",
            "example": "Follow-up visit for migraine symptoms"
          }
        },
        "required": [
          "doctorId",
          "hospitalId",
          "appointmentDate",
          "startTime"
        ]
      },
      "AppointmentResponse": {
        "type": "object",
        "description": "Detailed appointment response.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Appointment identifier.",
            "example": 501
          },
          "patientId": {
            "type": "integer",
            "format": "int64",
            "description": "Patient identifier.",
            "example": 201
          },
          "patientName": {
            "type": "string",
            "description": "Patient name.",
            "example": "Asha Patil"
          },
          "doctorId": {
            "type": "integer",
            "format": "int64",
            "description": "Doctor identifier.",
            "example": 301
          },
          "doctorName": {
            "type": "string",
            "description": "Doctor name.",
            "example": "Dr. Meera Shah"
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64",
            "description": "Hospital identifier.",
            "example": 12
          },
          "hospitalName": {
            "type": "string",
            "description": "Hospital name.",
            "example": "Sunrise Care Hospital"
          },
          "appointmentDate": {
            "type": "string",
            "format": "date",
            "description": "Appointment date.",
            "example": "2026-07-01"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Appointment start time.",
            "example": "10:30:00"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "Appointment end time.",
            "example": "10:45:00"
          },
          "status": {
            "type": "string",
            "description": "Current appointment status.",
            "example": "SCHEDULED"
          },
          "isOffline": {
            "type": "boolean",
            "description": "Whether the appointment is offline.",
            "example": false
          },
          "notes": {
            "type": "string",
            "description": "Appointment notes.",
            "example": "Follow-up visit for migraine symptoms"
          },
          "bookedByRole": {
            "type": "string",
            "description": "Role that booked the appointment.",
            "example": "PATIENT"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the appointment was created.",
            "example": "2026-06-27T11:30:00Z"
          }
        }
      },
...
      "AppointmentSummaryPagedResponse": {
        "type": "object",
        "description": "Paginated list of appointment summaries.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/AppointmentSummaryResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
      "SlotResponse": {
        "type": "object",
        "description": "Available appointment slot window.",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Slot start time.",
            "example": "10:30:00"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "Slot end time.",
            "example": "10:45:00"
          }
        }
      },
...
      "StatusUpdateRequest": {
        "type": "object",
        "description": "Request payload to update an appointment status.",
        "properties": {
          "status": {
            "type": "string",
            "description": "New appointment status.",
            "example": "CANCELLED"
          },
          "notes": {
            "type": "string",
            "description": "Optional status update notes.",
            "example": "Patient requested cancellation"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "DoctorResponse": {
        "type": "object",
        "description": "Doctor detail response.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Doctor identifier.",
            "example": 301
          },
          "name": {
            "type": "string",
            "description": "Doctor name.",
            "example": "Dr. Meera Shah"
          },
          "specialization": {
            "type": "string",
            "description": "Doctor specialization.",
            "example": "Neurology"
          },
          "education": {
            "type": "string",
            "description": "Doctor educational qualification.",
            "example": "MD, DM Neurology"
          },
          "phone": {
            "type": "string",
            "description": "Doctor mobile number.",
            "example": "9988776655"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Doctor email address.",
            "example": "meera.shah@example.com"
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64",
            "description": "Hospital identifier.",
            "example": 12
          },
          "hospitalName": {
            "type": "string",
            "description": "Hospital name.",
            "example": "Sunrise Care Hospital"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the doctor account is active.",
            "example": true
          }
        }
      },
      "DoctorSummaryResponse": {
        "type": "object",
        "description": "Compact doctor summary.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Doctor identifier.",
            "example": 301
          },
          "name": {
            "type": "string",
            "description": "Doctor name.",
            "example": "Dr. Meera Shah"
          },
          "specialization": {
            "type": "string",
            "description": "Doctor specialization.",
            "example": "Neurology"
          },
          "education": {
            "type": "string",
            "description": "Doctor educational qualification.",
            "example": "MD, DM Neurology"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the doctor account is active.",
            "example": true
          }
        }
      },
      "DoctorSummaryPagedResponse": {
        "type": "object",
        "description": "Paginated list of doctor summaries.",
        "properties": {
          "content": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DoctorSummaryResponse" }
          },
          "page": { "type": "integer", "example": 0 },
          "size": { "type": "integer", "example": 20 },
          "totalElements": { "type": "integer", "format": "int64", "example": 42 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
      "CalendarDayResponse": {
        "type": "object",
        "description": "Slot availability summary for a single calendar day.",
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "The calendar date.",
            "example": "2026-07-22"
          },
          "status": {
            "type": "string",
            "enum": ["NO_SCHEDULE", "AVAILABLE", "FULL"],
            "description": "Availability status for the day.",
            "example": "AVAILABLE"
          },
          "totalSlots": {
            "type": "integer",
            "description": "Total slots configured for the day.",
            "example": 14
          },
          "bookedSlots": {
            "type": "integer",
            "description": "Number of booked (non-cancelled) appointments.",
            "example": 6
          },
          "availableSlots": {
            "type": "integer",
            "description": "Remaining open slots.",
            "example": 8
          },
          "morningTotalSlots": {
            "type": "integer",
            "description": "Total morning slots (before 12:00).",
            "example": 6
          },
          "morningAvailableSlots": {
            "type": "integer",
            "description": "Available morning slots (before 12:00).",
            "example": 3
          },
          "eveningTotalSlots": {
            "type": "integer",
            "description": "Total evening slots (12:00 and after).",
            "example": 8
          },
          "eveningAvailableSlots": {
            "type": "integer",
            "description": "Available evening slots (12:00 and after).",
            "example": 5
          }
        }
      },
      "ScheduleRequest": {
        "type": "object",
        "description": "Request payload to save a doctor's weekly schedule.",
        "properties": {
          "dayOfWeek": {
            "type": "integer",
            "minimum": 0,
            "maximum": 6,
            "description": "Day of week index where 0 is Monday and 6 is Sunday.",
            "example": 0
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Schedule start time.",
            "example": "09:00:00"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "Schedule end time.",
            "example": "13:00:00"
          },
          "slotDurationMinutes": {
            "type": "integer",
            "minimum": 5,
            "description": "Duration of each slot in minutes.",
            "example": 15
          },
          "sessionType": {
            "type": "string",
            "enum": ["MORNING", "EVENING"],
            "description": "Session type: MORNING or EVENING.",
            "example": "MORNING"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime",
          "slotDurationMinutes",
          "sessionType"
        ]
      },
      "ScheduleResponse": {
        "type": "object",
        "description": "Doctor schedule entry response.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Schedule identifier.",
            "example": 901
          },
          "dayOfWeek": {
            "type": "integer",
            "description": "Day of week index where 0 is Monday and 6 is Sunday.",
            "example": 0
          },
          "dayLabel": {
            "type": "string",
            "description": "Human-readable day label.",
            "example": "Monday"
          },
          "startTime": {
            "type": "string",
            "format": "time",
            "description": "Schedule start time.",
            "example": "09:00:00"
          },
          "endTime": {
            "type": "string",
            "format": "time",
            "description": "Schedule end time.",
            "example": "13:00:00"
          },
          "slotDurationMinutes": {
            "type": "integer",
            "description": "Duration of each slot in minutes.",
            "example": 15
          },
          "sessionType": {
            "type": "string",
            "enum": ["MORNING", "EVENING"],
            "description": "Session type: MORNING or EVENING.",
            "example": "MORNING"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the schedule entry is active.",
            "example": true
          }
        }
      },
...
    }
  }
}